+++ internal/market/twill_gen.go
@@ -781,6 +781,7 @@ type __is_CreateRequest[T ~struct {
 	MerchantID    string   "json:\"merchant_id\""
 	EventID       string   "json:\"event_id\""
 	Type          string   "json:\"type\""
+	Category      string   "json:\"category,omitempty\""
 	Question      string   "json:\"question\""
 	Options       []string "json:\"options\""
 	LiquidityPool float64  "json:\"liquidity_pool\""
@@ -795,6 +796,7 @@ func (x *CreateRequest) WeaverMarshal(enc *codegen.Encoder) {
 	enc.String(x.MerchantID)
 	enc.String(x.EventID)
 	enc.String(x.Type)
+	enc.String(x.Category)
 	enc.String(x.Question)
 	servicetwill_enc_slice_string_4af10117(enc, x.Options)
 	enc.Float64(x.LiquidityPool)
@@ -807,6 +809,7 @@ func (x *CreateRequest) WeaverUnmarshal(dec *codegen.Decoder) {
 	x.MerchantID = dec.String()
 	x.EventID = dec.String()
 	x.Type = dec.String()
+	x.Category = dec.String()
 	x.Question = dec.String()
 	x.Options = servicetwill_dec_slice_string_4af10117(dec)
 	x.LiquidityPool = dec.Float64()
@@ -841,6 +844,7 @@ type __is_ListFilters[T ~struct {
 	twill.AutoMarshal
 	MerchantID string "json:\"merchant_id,omitempty\""
 	EventID    string "json:\"event_id,omitempty\""
+	Category   string "json:\"category,omitempty\""
 	Status     string "json:\"status,omitempty\""
 	Sort       string "json:\"sort,omitempty\""
 	Page       int    "json:\"page,omitempty\""
@@ -855,6 +859,7 @@ func (x *ListFilters) WeaverMarshal(enc *codegen.Encoder) {
 	}
 	enc.String(x.MerchantID)
 	enc.String(x.EventID)
+	enc.String(x.Category)
 	enc.String(x.Status)
 	enc.String(x.Sort)
 	enc.Int(x.Page)
@@ -867,6 +872,7 @@ func (x *ListFilters) WeaverUnmarshal(dec *codegen.Decoder) {
 	}
 	x.MerchantID = dec.String()
 	x.EventID = dec.String()
+	x.Category = dec.String()
 	x.Status = dec.String()
 	x.Sort = dec.String()
 	x.Page = dec.Int()
@@ -1073,6 +1079,7 @@ func servicetwill_size_ListFilters_f7a95a88(x *ListFilters) int {
 	size += 0
 	size += (4 + len(x.MerchantID))
 	size += (4 + len(x.EventID))
+	size += (4 + len(x.Category))
 	size += (4 + len(x.Status))
 	size += (4 + len(x.Sort))
 	size += 8